Extracted TimeZoneSearch and ContactList Into Their Own Components
